refactor: add a coin list component
